train/requirements.txt
# Training dependencies for BLUX-cA QLoRA pipeline
accelerate>=0.31.0
transformers>=4.41.0
peft>=0.11.0
trl>=0.9.4
bitsandbytes>=0.43.0
datasets>=2.19.0
pyyaml>=6.0.1
tqdm>=4.66.0
